requirements-p4.txt
PyYAML>=6.0
dash>=2.18
plotly>=5.24
torch>=2.2
torchvision>=0.17
transformers>=4.39
sentencepiece>=0.1.99
accelerate>=0.28
